[refactor][server] Split server into modular function-based modules
Split the 1071-line __main__.py into focused modules:
- log.py: Shared _log helper (22 LOC)
- notify.py: Telegram message formatting and sending (156 LOC)
- session.py: Session state file lookups (91 LOC)
- github.py: Issue/PR discovery and GraphQL helpers (394 LOC)
- workers.py: Worktree spawn/rebase and worker status (229 LOC)
- __main__.py: CLI entry point and coordinator (258 LOC)

All public functions are re-exported from __main__.py to preserve
backward compatibility with existing test imports.

Closes #371
